M Daemon

+++ mail/m-daemon.sh
@@ -2,20 +2,24 @@
 
 kilrek()
 {
-	pgrep -P $1 | while read p;
-	do
-		echo $1 $2 to $p;
-		if [[ "$p" != "$2" ]]
-		then
-			kilrek $p $2
-			echo $p;
-			kill $p;
-		fi
-	done
+	if (( $3 > 0 ));
+	then
+		pgrep -P $1 | while read p;
+		do
+			# echo $1 $2 to $p;
+			if [[ "$p" != "$2" ]]
+			then
+				kilrek $p $2 $(( $3 - 1 ))
+				# echo $p;
+				kill $p;
+			fi
+		done
+	fi
 
 }
 whileok()
 {
+	(
 	while true
 	do
 		sleep 60
@@ -23,15 +27,14 @@ whileok()
 		delay=$(( $(date +%s) - ${x[0]} ))
 		if (( $delay > 120 ))
 		then
-			(
-			echo KILL
+			# echo KILL
 			pidwok=$BASHPID
-			echo pidwok $pidwok
-			kilrek $1 $pidwok
+			# echo pidwok $pidwok
+			kilrek $1 $pidwok 5
 			return;
-		)
 		fi
 	done
+)
 }
 
 pid=$BASHPID
@@ -41,6 +44,10 @@ do
 	echo -ne "\033]0;MAIL\007"
 	osdc --log=2 --color=red "TRY START MAIL"
 	m
+	(
+		pidwok=$BASHPID
+		kilrek $pid $pidwok 5
+	)
 	sleep 2
 done
